+++ apps/repos/urls.py
@@ -6,5 +6,6 @@ app_name = "repos"
 urlpatterns = [
     path("repos/", views.RepositorySearchView.as_view(), name="search"),
     path("repos/<str:owner>/<str:name>/", views.RepositoryDetailView.as_view(), name="repo_detail"),
+    path("lists/", views.AwesomeListListView.as_view(), name="list"),
     path("lists/<slug:slug>/", views.AwesomeListDetailView.as_view(), name="list_detail"),
 ]